--- taskmanager-web/src/main/web/src/app/viewTask/viewtask-routing.module.ts
@@ -1,13 +1,0 @@
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-import { ViewTaskComponent } from './viewtask.component';
-
-@NgModule({
-  imports: [
-    RouterModule.forChild([
-      { path: 'viewtask', component: ViewTaskComponent }
-    ])
-  ],
-  exports: [RouterModule]
-})
-export class ViewTaskRoutingModule { }
