--- taskmanager-web/src/main/web/src/app/viewTask/viewtask.component.css
@@ -1,4 +1,0 @@
-th, label, .h6{
-    font-weight: 500 !important;
-    color: #495057 !important;
-}
