+++ taskmanager-web/src/main/web/src/app/viewtasklist/viewtask.component.css
@@ -1,0 +1,5 @@
+th, label, .h6{
+    font-weight: 500 !important;
+    color: #495057 !important;
+}
+
